logo-ann/generation/requirements_streamlit.txt
annoy
Pillow
requests
streamlit
